[APP-6968] Review fixes: cap non-string payloads, float costs, lowercase trace ids
- truncation: count, dedup and window container attribute values (OTLP
  arrays, JSON-decoded message lists) over their serialized JSON via new
  cap_payload_value/_payload_text helpers — counting only str values let a
  100k-char gen_ai.prompt list bypass every cap and blow the span budget
- client: min/max_trace_cost typed float | None (trace cost is a fractional
  currency amount; int rejected every realistic sub-dollar bound)
- utils: require_trace_id returns the id lowercased (the server emits and
  matches lowercase hex)

+++ src/datarobot_genai/drtools/core/clients/datarobot_otel_query.py
@@ -64,8 +64,8 @@ class OtelQueryApiClient:
         min_trace_duration_ns: int | None = None,
         min_span_duration_ns: int | None = None,
         max_span_duration_ns: int | None = None,
-        min_trace_cost: int | None = None,
-        max_trace_cost: int | None = None,
+        min_trace_cost: float | None = None,
+        max_trace_cost: float | None = None,
         sort_by: str | None = None,
         sort_direction: str | None = None,
     ) -> dict[str, Any]:
